Relocate `get_num_xcd` into the dependency-free Triton launcher
`get_num_xcd` (AMD CDNA XCD count for `xcd_remap`) is runtime grid logic that
generated code invokes as `helion.runtime.get_num_xcd(...)`, so it belongs with
the other dependency-free launch helpers. Move it (and `_CUS_PER_XCD`) from
`helion/_compat.py` into `helion/runtime/triton/launcher.py`, making the launcher
the complete dependency-free home the precompiler bulk-exports.

It is re-exported unchanged from `helion.runtime` (so the generated
`helion.runtime.get_num_xcd(...)` still resolves); the two importers move to the
new location (`autotuner/config_spec.py`, `_testing.py`). `get_num_xcd` stays
torch-only, so no import cycle is introduced -- `config_spec` ->
`helion.runtime.triton.launcher` is acyclic because the launcher imports no
`helion` module.

Verified on H100: `import helion` clean, `get_num_xcd` is identical from all
re-export sites, `ruff`/`pyrefly` clean (83 baseline), `test_config_api` /
`test_runtime` pass.

stack-info: PR: https://github.com/pytorch/helion/pull/3180, branch: AmesingFlank/stack/94
